R6 class containing computed freqtable

Computed frequency table class. On basis of raw scores of some variables, it creates a frequency table and allows to compute scores in some standard scale. To initialize the object of this class the 'gen_freqtable' function is recommended.

Methods

Public methods

Comp_Freqtable$get_status()

Comp_Freqtable$get_data_info()

Comp_Freqtable$new()

Comp_Freqtable$get_freqtables()

Comp_Freqtable$attach_scale()

Comp_Freqtable$compute_scores()

Comp_Freqtable$get_scoretables()

Comp_Freqtable$get_computed_scores()

Comp_Freqtable$get_computed_scores_ext()

Comp_Freqtable$clone()

Method get_status()

Current information about object

Usage

Comp_Freqtable$get_status()

Returns

List of metrics

n: Number of observations for frequency table generation

kept_data: Is the data kept in the object?

frequency_tables: Are the frequency table complete or incomplete?

standardized_scores: Are the standardized scores computed - and if yes, then: for which scales?

Method get_data_info()

Information about kept data

Usage

Comp_Freqtable$get_data_info()

Method new()

Initialization of the 'comp_freqtable' class

Usage

Comp_Freqtable$new(data, vars, id, keep_data = T)

Arguments

data

data to use for computing.

vars

character vector containing column names of numeric variables that you want to calculate frequency tables for.

id

character string containing column name for identification of the observations. Needed only if 'keep_data' is TRUE

keep_data

do you want to keep the table for getting scores for individual observations? Defaults to TRUE.

Method get_freqtables()

Get generated frequency tables.

Usage

Comp_Freqtable$get_freqtables()

Returns

Two lists, containing actual cumulated frequency tables and their statuses

freqtables: Cumulated frequency tables for each variable specified during initialization

statuses: Statuses for each frequency table. 'complete' if there was at least one observation for each raw score between actual minimum and maximum. 'incomplete' if there was at least one missing raw score and they were filled automatically

Method attach_scale()

Attach new scale to the object

Usage

Comp_Freqtable$attach_scale(scale)

Arguments

scale

FreqtableScale object defining a scale

Method compute_scores()

Compute scores in scale of your choice. After computation, you can get them with 'get_scoretables' method.

Usage

Comp_Freqtable$compute_scores(scale)

Arguments

scale

Either a character string indicating one of built-in scale, or FreqtableScale object describing new scale. If new scale is declared, it will be automatically attached

Method get_scoretables()

Get computed scoring tables for all variables

Usage

Comp_Freqtable$get_scoretables(scale)

Arguments

scale

character string indicating for which computed scale return values

Returns

List containing the actual tables with raw score and standardized score pairs

Method get_computed_scores()

Get computed scores for specified ids or all sourced observations. You can only use this method if you chose to keep source data.

Usage

Comp_Freqtable$get_computed_scores(scale, vars = NULL, ids = NULL)

Arguments

scale

character string defining the computed scale. Mandatory argument.

vars

character vector declaring for which variables of the embedded data to get the standardized score. If not specified, returns values for all

ids

for the observations in kept data. If none is provided, then it is returned for all observations.

Method get_computed_scores_ext()

Get computed scores for external observations

Usage

Comp_Freqtable$get_computed_scores_ext(data, scale, vars, attach = TRUE)

Arguments

data

data.frame containing raw scores to get standardized scores for

scale

character string defining the computed scale

vars

character vector declaring variables for which to get standardized score. If the variables in external data are named differently, the vector should be named accordingly.

attach

boolean indicating if new data should be attached to the frequency table. Defaults to TRUE.

Method clone()

The objects of this class are cloneable with this method.

Usage

Comp_Freqtable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.