Normalize raw scores

Use computed FrequencyTable or ScoreTable to normalize the provided raw scores.

normalize_score(x, table, what)

Arguments

x

vector of raw scores to normalize

table

FrequencyTable or ScoreTable object

what

the values to get. One of either:

quan - the quantile of x in the raw score distribution

Z - normalized Z score for the x raw score

name of the scale calculated in StandardScale provided to x argument

Value

Numeric vector with values specified in what argument

Examples

# normalize with FrequencyTable
ft <- FrequencyTable(HEXACO_60$HEX_H)
#> Warning: There are missing score values between minimum and maximum scores. They have been filled automatically.

normalizeScore(HEXACO_60$HEX_H[1:5], ft)
#> Error in normalizeScore(HEXACO_60$HEX_H[1:5], ft): could not find function "normalizeScore"

# normalize with ScoreTable
st <- ScoreTable(ft, list(STEN, STANINE))

normalizeScore(HEXACO_60$HEX_H[1:5], st, "sten")
#> Error in normalizeScore(HEXACO_60$HEX_H[1:5], st, "sten"): could not find function "normalizeScore"
normalizeScore(HEXACO_60$HEX_H[1:5], st, "stanine")
#> Error in normalizeScore(HEXACO_60$HEX_H[1:5], st, "stanine"): could not find function "normalizeScore"