Gerenic plot of the GroupedFrequencyTable

Generic plot using ggplot2. It plots FrequencyTables for all groups by default, or only chosen ones using when group_names argument is specified.

Usage

# S3 method for GroupedFrequencyTable
plot(
  x,
  group_names = NULL,
  strict_names = TRUE,
  plot_grid = is.intersected(x),
  ...
)

Arguments

x

A GroupedFrequencyTable object

group_names

vector specifying which groups should appear in the plots

strict_names

If TRUE, then intersected groups are filtered using strict strategy: group_names need to be provided in form: "group1:group2". If FALSE, then intersected groups will be taken into regard separately, so eg. when "group1" is provided to group_names, all of: "group1:group2", "group1:group3", "group1:groupN" will be plotted. Defaults to TRUE

plot_grid

boolean indicating if the ggplot2::facet_grid() should be used. If FALSE, then ggplot2::facet_wrap() is used. If groups are not intersected, then it will be ignored and facet_wrap will be used.

...

named list of additional arguments passed to facet function used.