Create a ScoreTable

Creates a table to calculate scores in specified standardized scale for each discrete raw score. Uses normalization provided by FrequencyTable and scale definition created with StandardScale.

After creation it can be used to normalize and standardize raw scores with normalize_score.

plot.ScoreTable method requires ggplot2 package to be installed.

Usage

ScoreTable(ft, scale)

# S3 method for ScoreTable
print(st, max = NULL)

# S3 method for ScoreTable
plot(st, scale_name = NULL)

Arguments

ft

a FrequencyTable object

scale

a StandardScale object or list of multiple StandardScale objects

st

a ScoreTable object

max

numeric or NULL, specifying the maximal number of entries to be printed. By default, when NULL, getOption("max.print") used.

scale_name

if scores for multiple scales available, provide the name of the scale for plotting.

Value

object of class ScoreTable. Consists of:

table: data.frame containing for each point in the raw score:

number of observations (n),

frequency in sample (freq),

quantile (quan),

normalized Z-score (Z),

score transformed to every of provided StandardScales

status: list containing the total number of simulated observations (n) and information about raw scores range completion (range): complete or incomplete

scale: named list of all attached StandardScale objects

Examples

# firstly compute FrequencyTable for a variable
ft <- FrequencyTable(HEXACO_60$HEX_A)
#> Warning: There are missing score values between minimum and maximum scores. They have been filled automatically.

# then create a ScoreTable
st <- ScoreTable(ft, STEN)

# ScoreTable is ready to use!
st
#> <ScoreTable> computed on: 204 observations
#> 
#>  score  n      freq       quan           Z sten
#>     12  1 0.4901961  0.2450980 -2.81340671    1
#>     13  1 0.4901961  0.7352941 -2.43954226    1
#>     14  0 0.0000000  0.7352941 -2.43954226    1
#>     15  2 0.9803922  1.4705882 -2.17792307    1
#>     16  1 0.4901961  2.2058824 -2.01297132    1
#>     17  3 1.4705882  3.1862745 -1.85409556    2
#>     18  4 1.9607843  4.9019608 -1.65443473    2
#>     19  4 1.9607843  6.8627451 -1.48609157    3
#>     20  3 1.4705882  8.5784314 -1.36718085    3
#>     21  7 3.4313725 11.0294118 -1.22496546    3
#>     22 11 5.3921569 15.4411765 -1.01769373    3
#>     23 11 5.3921569 20.8333333 -0.81221780    4
#>     24  9 4.4117647 25.7352941 -0.65152773    4
#>     25  5 2.4509804 29.1666667 -0.54852228    4
#>     26  7 3.4313725 32.1078431 -0.46468526    5
#>     27  8 3.9215686 35.7843137 -0.36422999    5
#>     28 13 6.3725490 40.9313725 -0.22931068    5
#>     29 13 6.3725490 47.3039216 -0.06763219    5
#>     30 16 7.8431373 54.4117647  0.11081291    6
#>     31 14 6.8627451 61.7647059  0.29930691    6
#>     32  9 4.4117647 67.4019608  0.45103991    6
#>     33 13 6.3725490 72.7941176  0.60659812    7
#>     34 13 6.3725490 79.1666667  0.81221780    7
#>     35  7 3.4313725 84.0686275  0.99728241    7
#>     36  5 2.4509804 87.0098039  1.12685469    8
#>     37  4 1.9607843 89.2156863  1.23808034    8
#>     38  7 3.4313725 91.9117647  1.39916101    8
#>     39  3 1.4705882 94.3627451  1.58597448    9
#>     40  3 1.4705882 95.8333333  1.73166440    9
#>     41  3 1.4705882 97.3039216  1.92746610    9
#>     42  2 0.9803922 98.5294118  2.17792307   10
#>     43  0 0.0000000 98.5294118  2.17792307   10
#>     44  1 0.4901961 99.2647059  2.43954226   10
#>     45  1 0.4901961 99.7549020  2.81340671   10
#> 
#> 
#> Used StandardScales:
#> <StandardScale>: 'sten'
#> ( M: 5.5; SD: 2; min: 1; max: 10 )
#>