Normalize raw scores

Use computed FrequencyTable or ScoreTable to normalize the provided raw scores.

Usage

normalize_score(x, table, what)

Arguments

x

vector of raw scores to normalize

table

FrequencyTable or ScoreTable object

what

the values to get. One of either:

quan - the quantile of x in the raw score distribution

Z - normalized Z score for the x raw score

name of the scale calculated in StandardScale provided to x argument

Value

Numeric vector with values specified in what argument

See also

Examples

# normalize with FrequencyTable
ft <- FrequencyTable(HEXACO_60$HEX_H)
#> Warning: There are missing score values between minimum and maximum scores. They have been filled automatically.

normalize_score(HEXACO_60$HEX_H[1:5], ft)
#> Error in what %in% c("quan", "Z"): argument "what" is missing, with no default

# normalize with ScoreTable
st <- ScoreTable(ft, list(STEN, STANINE))

normalize_score(HEXACO_60$HEX_H[1:5], st, "sten")
#> [1] 8 7 2 2 5
normalize_score(HEXACO_60$HEX_H[1:5], st, "stanine")
#> [1] 7 6 1 2 4