Normalize scores using GroupedFrequencyTables or GroupedScoreTables

Normalize scores using either GroupedFrequencyTable or GroupedScoreTable for one or more variables. Given data.frame should also contain columns used in GroupingConditions attached to the table

Usage

normalize_scores_grouped(
  data,
  vars,
  ...,
  what,
  retain = FALSE,
  group_col = NULL,
  .dots = list()
)

Arguments

data

data.frame object containing raw scores

vars

names of columns to normalize. Length of vars need to be the same as number of tables provided to either ... or .dots

...

GroupedFrequencyTable or GroupedScoreTable objects to be used for normalization. They should be provided in the same order as vars

what

the values to get. One of either:

quan - the quantile of x in the raw score distribution

Z - normalized Z score for the x raw score

name of the scale calculated in GroupedScoreTables provided to ... or .dots argument

retain

either boolean: TRUE if all columns in the data are to be retained, FALSE if none; or character vector with names of columns to be retained

group_col

name of the column for name of the group each observation was qualified into. If left as default NULL, they won't be returned.

.dots

GroupedFrequencyTable or GroupedScoreTable objects provided as a list, instead of individually in ....

Value

data.frame with normalized scores

See also

Examples

#### create GroupConditions for calculating tables ####
age_grouping <- GroupConditions(
  conditions_category = "Age",
  "below 18" ~ age < 18,
  "18-22" ~ age >= 18 & age <= 22,
  "23-26" ~ age >= 23 & age <= 26,
  "27-32" ~ age >= 27 & age <= 32,
  "33-40" ~ age >= 33 & age <= 40,
  "40-60" ~ age >= 40 & age <= 60,
  "above 60" ~ age > 60
)
sex_grouping <- GroupConditions(
  conditions_category = "Sex",
  "Male" ~ sex == "M",
  "Female" ~ sex == "F"
)

#### create GroupedFrequencyTable ####
NEU_gft <- GroupedFrequencyTable(
  data = IPIP_NEO_300,
  conditions = list(age_grouping, sex_grouping),
  var = "N"
)
#> ℹ There are missing raw score values between minimum and maximum raw scores for
#>   some groups. They have been filled automatically.
#> • below 18:Male No. missing: 34/202; 16.83%
#> • below 18:Female No. missing: 39/195; 20%
#> • below 18:.all2 No. missing: 30/214; 14.02%
#> • 18-22:Male No. missing: 37/213; 17.37%
#> • 18-22:Female No. missing: 40/202; 19.8%
#> • 18-22:.all2 No. missing: 37/230; 16.09%
#> • 23-26:Male No. missing: 38/211; 18.01%
#> • 23-26:Female No. missing: 52/218; 23.85%
#> • 23-26:.all2 No. missing: 32/225; 14.22%
#> • 27-32:Male No. missing: 47/220; 21.36%
#> • 27-32:Female No. missing: 42/208; 20.19%
#> • 27-32:.all2 No. missing: 26/220; 11.82%
#> • 33-40:Male No. missing: 26/206; 12.62%
#> • 33-40:Female No. missing: 41/208; 19.71%
#> • 33-40:.all2 No. missing: 25/221; 11.31%
#> • 40-60:Male No. missing: 24/203; 11.82%
#> • 40-60:Female No. missing: 19/194; 9.79%
#> • 40-60:.all2 No. missing: 12/205; 5.85%
#> • above 60:Male No. missing: 60/208; 28.85%
#> • above 60:Female No. missing: 55/199; 27.64%
#> • above 60:.all2 No. missing: 39/209; 18.66%
#> • .all1:Male No. missing: 6/221; 2.71%
#> • .all1:Female No. missing: 12/220; 5.45%
#> • .all1:.all2 No. missing: 6/230; 2.61%

#### create GroupedScoreTable ####
NEU_gst <- GroupedScoreTable(
  NEU_gft,
  scale = list(STEN, STANINE)
)

#### normalize scores ####
# to Z score or quantile using GroupedFrequencyTable
normalized_to_quan <- normalize_scores_grouped(
  IPIP_NEO_300,
  vars = "N",
  NEU_gft,
  what = "quan",
  retain = c("sex", "age")
)

# only 'sex' and 'age' are retained
head(normalized_to_quan)
#>    sex age        N
#> 1:   F  25 57.72613
#> 2:   F  18 79.69349
#> 3:   F  16 75.96996
#> 4:   M  23 59.94994
#> 5:   F  25 33.29146
#> 6:   M  26 80.60075

# to StandardScale attached to GroupedScoreTable
normalized_to_STEN <- normalize_scores_grouped(
  IPIP_NEO_300,
  vars = "N",
  NEU_gst,
  what = "stanine",
  retain = FALSE,
  group_col = "sex_age_group"
)

# none is retained, 'sex_age_goup' is created
head(normalized_to_STEN)
#>      sex_age_group N
#> 1:    23-26:Female 5
#> 2:    18-22:Female 7
#> 3: below 18:Female 6
#> 4:      23-26:Male 6
#> 5:    23-26:Female 4
#> 6:      23-26:Male 7